+++ src/ApiConnector.php
@@ -8,10 +8,13 @@ use Atst\Guzzle\Http\Plugin\WsseAuthPlugin;
 use Psr\Log\LoggerInterface;
 use Psr\Log\NullLogger;
 
-use Penneo\SDK\Entity;
-
 class ApiConnector
 {
+    /**
+     * We should always keep this up to date. The '>=' is to allow for human error.
+     */
+    const VERSION = '>=v1.15.0';
+
     static protected $endpoint;
     static protected $headers;
     static protected $lastError;
@@ -45,10 +48,11 @@ class ApiConnector
     {
         self::$endpoint = $endpoint ?: self::getDefaultEndpoint();
 
-        self::$headers = self::getDefaultHeaders();
-        if ($headers) {
-            self::$headers = array_merge($headers, self::$headers);
-        }
+        self::$headers = array_merge(
+            $headers ?: [],
+            self::getDefaultHeaders(),
+            self::getSpecificHeaders($key, $headers)
+        );
 
         if ($user) {
             self::$headers['penneo-api-user'] = (int) $user;
@@ -136,7 +140,9 @@ class ApiConnector
             $request = self::$client->createRequest($method, $url, self::$headers, $data, $options);
             $response = $request->send();
             if ($response instanceof Response) {
-                self::$logger->debug('response', [
+                // some logging implementation might not print the context, we put the request id in the log message
+                // because it is important and we want to make sure it gets seen
+                self::$logger->debug('response requestId=' . $response->getHeader('X-Penneo-Request-Id'), [
                     'method' => $method,
                     'url'    => $url,
                     'raw'    => $response->getMessage()
@@ -148,7 +154,7 @@ class ApiConnector
             $response = $request->getResponse();
             if ($response instanceof Response) {
                 $message = $response->getMessage();
-                self::$logger->error('response', [
+                self::$logger->error('response requestId=' . $response->getHeader('X-Penneo-Request-Id'), [
                     'method' => $method,
                     'url'    => $url,
                     'raw'    => $message
@@ -167,5 +173,24 @@ class ApiConnector
     public static function getLastError()
     {
         return self::$lastError;
+    }
+
+    /**
+     * @param string     $key
+     * @param array|null $headers
+     *
+     * @return array<string, string>
+     */
+    private static function getSpecificHeaders($key, array $headers = null)
+    {
+        $keyPart = substr($key, 0, 8);
+        $setUserAgent = $headers && array_key_exists('User-Agent', $headers) ?
+            $headers['User-Agent'] : 'n/a';
+        $version = self::VERSION;
+
+        return [
+            // this helps us identify API users if we spot incorrect usage of the API or if we discover potential errors
+            'User-Agent' => "penneo/penneo-sdk-php v:${version} key:${keyPart} ua:${setUserAgent}"
+        ];
     }
 }