+++ src/CaseFile.php
@@ -18,6 +18,7 @@ class CaseFile extends Entity
             'caseFileTypeId' => 'caseFileType->getId',
             'sensitiveData',
             'disableNotificationsOwner',
+            'disableEmailAttachments',
             'reference',
         ),
         'update' => array(
@@ -28,29 +29,49 @@ class CaseFile extends Entity
             'visibilityMode',
             'documentDisplayMode',
             'disableNotificationsOwner',
+            'disableEmailAttachments',
             'sensitiveData',
             'reference',
         )
     );
     protected static $relativeUrl = 'casefiles';
 
+    /** @var string */
     protected $title;
+    /** @var string */
     protected $language;
+    /** @var string */
     protected $metaData;
+    /** @var int */
     protected $sendAt;
+    /** @var int */
     protected $expireAt;
+    /** @var int */
     protected $visibilityMode;
+    /** @var int */
     protected $documentDisplayMode;
+    /** @var string */
     protected $reference;
+    /** @var bool */
     protected $sensitiveData;
+    /** @var bool */
     protected $disableNotificationsOwner;
+    /** @var bool */
+    protected $disableEmailAttachments = false;
+    /** @var int */
     protected $status;
+    /** @var int */
     protected $created;
+    /** @var int */
     protected $signIteration;
+    /** @var CaseFileTemplate|null */
     protected $caseFileType;
 
+    /** @var Document[]|null */
     protected $documents = null;
+    /** @var Signer[]|null */
     protected $signers = null;
+    /** @var CopyRecipient[]|null */
     protected $copyRecipients = null;
 
     public function __construct()
@@ -311,6 +332,39 @@ class CaseFile extends Entity
     public function getDisableNotificationsOwner()
     {
         return $this->disableNotificationsOwner;
+    }
+
+    /**
+     * With this setting on, emails sent to signers and the case file owner when the case file gets finalized
+     * will not have the signed PDF files attached.
+     *
+     * Using this feature together with CopyRecipients is not support, the case file owner will receive warning emails
+     * stating that Copy Recipient emails could not be sent.
+     *
+     * Warning: You can only read this value if using the '/v2' and up versions of the base API URL.
+     *
+     * @return bool
+     */
+    public function getDisableEmailAttachments()
+    {
+        return $this->disableEmailAttachments;
+    }
+
+    /**
+     * With this setting on, emails sent to signers and the case file owner when the case file gets finalized
+     * will not have the signed PDF files attached.
+     *
+     * Using this feature together with CopyRecipients is not support, the case file owner will receive warning emails
+     * stating that Copy Recipient emails could not be sent.
+     *
+     * @param bool $disableEmailAttachments
+     * @return CaseFile
+     */
+    public function setDisableEmailAttachments($disableEmailAttachments)
+    {
+        $this->disableEmailAttachments = $disableEmailAttachments;
+
+        return $this;
     }
 
     public function getStatus()